requirements.txt
# 增强新闻分析系统依赖包

# 核心框架
pyspark==3.5.0
kafka-python==2.0.2
mysql-connector-python==8.2.0

# 数据处理
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# 自然语言处理
textblob==0.17.1
nltk==3.8.1

# 可视化
matplotlib==3.8.2
seaborn==0.13.0

# 机器学习扩展
scipy==1.11.4
joblib==1.3.2

# 实用工具
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# 开发工具
pytest==7.4.3
jupyter==1.0.0

# 性能优化
psutil==5.9.6